added CSS to grades page
