--- app/content/gradetableview.php
@@ -1,6 +1,0 @@
-<link rel="stylesheet" href="./CSS/all.css" type="text/css">
-<?php
-	function gradeTableView(){
-		echo "";
-		}
-?>
